Renaming file + placement of a button
