+++ react-components/src/components/Architecture/RevealButtons.tsx
